Updates to API calls, tabContents
Reworked the API calls now that data for traits, subraces and proficiencies is coming in correctly from the main /races endpoint call. Also added some basic error handling, and created variables for two variations of the api URL to reduce repetition. Changed tabContents so that they now only do their API calls when the tab is actually clicked on, however need to refine for cases where the user clicks to a different tab and then back to original one (it should store the data instead of having to do api calls each time)